+++ src/main/kotlin/de/mirb/project/miftp/control/FtpHandler.kt
@@ -7,9 +7,10 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 import java.nio.ByteBuffer
 import java.util.*
+import kotlin.collections.HashMap
 
 @Component
-class FtpHandler @Autowired constructor(private val server: MiFtpServer) {
+class FtpHandler @Autowired constructor(private val server: MiFtpServer, private val fileStore: FtpFileStore) {
 
   fun listFiles(user: String):List<FileEndpoint> {
     val view = server.getFileSystemView(user)
@@ -65,11 +66,13 @@ class FtpHandler @Autowired constructor(private val server: MiFtpServer) {
             .firstOrNull())
   }
 
+  fun getFileByToken(token: String): Optional<FileEndpoint> {
+    return fileStore.getFileByToken(token).map { FileEndpoint.create(it) }
+  }
+
   /**
    * Get current count of files (excluded directories)
    */
   fun getFilesCount(user: String) = listFiles(user).size
-
-//  fun list
 }